refactor(llm): extract the stream fallback engine from its vision binding
Pure move + rename. visionStreamFallback re-exports under the old names with
the same tuning, so the live vision path is behaviourally identical and its
tests are untouched. Direct Assist needs the same engine and must not share
the vision health map.

Also updates textStreamFallback.ts's DEFAULT_TEXT_FALLBACK_CONFIG with the
new required logPrefix field (set to 'Vision' to keep its existing, if
mislabeled, log text byte-identical) — a third pre-existing consumer of this
module the plan didn't account for; see task-1-report.md.
